pyproject.toml
[project]
name = "noise-detector"
version = "0.9.0"
description = "Very simple app to detect noise while human sleeping"
authors = [
    {name = "andrews-lerk", email = "j_k7@bk.ru"},
]
dependencies = [
    "pyaudio>=0.2.14",
    "colorlog>=6.8.2",
    "pyyaml>=6.0.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true
